Fix cold-temperature charge protection (audit findings #1, #2)
#1: find_cutback() compared a uint32_t threshold against a signed int
measurement, forcing negative temperatures to wrap to a huge unsigned
value. A sub-zero pack temp matched the highest HOT_CUTBACK entry and
clamped charge power to 0 W. Cast threshold to int so the comparison is
signed; all thresholds are small positive values, so the cast is lossless.

#2: COLD_CUTBACK was defined but never applied, so the firmware allowed
full power at low pack temperature (lithium-plating risk). rampTask now
reads monolithMinTemp and folds a cold-cutback power limit into powerLimit
alongside the voltage and hot limits (CC phase only), keyed off the
coldest thermocouple. Ramp log gains a COLD annotation when the cold limit
actually constrains power.

The two fixes are coupled: cold protection only behaves correctly for
negative temps because of the signedness fix in #1.

+++ development/battery_tables.h
@@ -18,15 +18,21 @@ uint32_t find_cutback(
   const cutback_entry *entries,
   size_t entry_count
 ) {
+  // `threshold` is uint32_t but `measurement` is a signed int (temperatures
+  // can be negative). Compare as signed: an unsigned comparison would convert
+  // a negative measurement to a huge value, so e.g. a sub-zero pack temp would
+  // match the highest HOT_CUTBACK entry and clamp charge power to 0. All
+  // thresholds in the tables are small positive values, so the (int) cast is
+  // lossless.
   if (mode == CUTBACK_AT_OR_ABOVE) {
     for (ssize_t i = entry_count - 1; i >= 0; i--) {
-      if (entries[i].threshold <= measurement) {
+      if ((int)entries[i].threshold <= measurement) {
         return entries[i].c_rate_thousandths;
       }
     }
   } else {
     for (ssize_t i = 0; i < entry_count; i++) {
-      if (entries[i].threshold >= measurement) {
+      if ((int)entries[i].threshold >= measurement) {
         return entries[i].c_rate_thousandths;
       }
     }
